Jag har uppdaterat alla kommenterar, fixat så att det inte är för många kommenterar, ändrat namn på Batterinivå till BatteriNivå, samt ändrat vissa this.hide() till this.Close() för mer korrekt beskrivning och metod för vad som faktiskt sker. Har dessutom lagt in regioner för de lite längre kodsnuttarna. samt tagit bort en till StationRepository dublett som jag hittade.

+++ LogicLayer/UthyrningsHistorik.cs
@@ -12,8 +12,8 @@ namespace BusinessEntities
         public DateTime SlutTid { get; private set; }
 
 
-        public int PrisPerMinut { get; private set; }
-        public UthyrningsHistorik(User user, DateTime startTid, DateTime slutTid, int prisPerMinut, Fordon fordon) //Konstruktor för UthyrningsHistorik
+        public double PrisPerMinut { get; private set; }
+        public UthyrningsHistorik(User user, DateTime startTid, DateTime slutTid, double prisPerMinut, Fordon fordon) //Konstruktor för UthyrningsHistorik
         {
             User = user;
             StartTid = startTid;
